Reorganised code so that the gpu code lives in its own folder. Changed imports from the models using GPU features accordingly.
