+++ topo/submodel/gcal.py
@@ -12,6 +12,7 @@ import topo.transferfn.misc
 from topo.submodel import Model, ArraySpec, order_projections # pyflakes:ignore (API import)
 from topo.submodel.earlyvision import EarlyVisionModel
 
+
 @Model.definition
 class ModelGCAL(EarlyVisionModel):
